# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mahfil Fund API",
    "version": "2.0.0",
    "description": "API-owned authentication and path-scoped multi-tenant operations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Login": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CompleteFirstLogin": {
        "type": "object",
        "required": [
          "challengeToken",
          "newPassword"
        ],
        "properties": {
          "challengeToken": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 10
          }
        }
      },
      "MemberCreate": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind",
              "email",
              "fullName",
              "temporaryPassword",
              "role"
            ],
            "properties": {
              "kind": {
                "const": "new"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "fullName": {
                "type": "string"
              },
              "temporaryPassword": {
                "type": "string",
                "minLength": 10
              },
              "role": {
                "enum": [
                  "admin",
                  "collector",
                  "viewer"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "kind",
              "email",
              "role"
            ],
            "properties": {
              "kind": {
                "const": "existing"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "role": {
                "enum": [
                  "admin",
                  "collector",
                  "viewer"
                ]
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session or password-change challenge"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/complete-first-login": {
      "post": {
        "operationId": "completeFirstLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteFirstLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created"
          }
        }
      }
    },
...
    "/communities/{communityId}/members": {
      "parameters": [
        {
          "name": "communityId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listMembers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Community members"
          }
        }
      },
      "post": {
        "operationId": "createMember",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account and membership created"
          }
        }
      }
    },
...
  }
}